requirements.txt
google-cloud-pubsub
google-cloud-bigquery
apache-beam
faker
pandas
psycopg2